RESTful_Web_API

+++ src/Cart/Cart.Api/Core/Models/Item.cs
@@ -2,10 +2,10 @@
 {
     public class Item
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public Image Image { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
     }
 }
